Add quick-fix to resolve unresolved references by adding an import statement
When the `UnresolvedReferenceAnnotator` finds a reference that it cannot
resolve, it will now provide a 'Add Import' quick-fix to attempt to
resolve the reference by adding an import statement for a module containing
the unresolved name.

In the case where a single candidate is found, the import will be created
silently. If multiple candidates are found, the user will be shown a popup
list of modules from which the name can be resolved. If there are no
candidates, an error hint will be shown.